Removed extra clone operation inside Trajopt.
Previously, due to bugs in the openrave environment cloning,
it was necessary to re-clone a fresh environment for TrajOpt on
each planning call to avoid stale OSG and Bullet references
corrupting the TrajOpt planner.

With new fixes, it is possible to simply exhaustively clear the
TrajOpt userdata on each planning call, removing this stale data
at much less cost than fully re-cloning the environment.